Разделил Gulp на задачи и добавил импорты с помощью Babel, добавил EsLint и Prettier
