+++ .gitignore
@@ -1,3 +1,4 @@
 /node_modules
 /app
+/.idea
 package-lock.json
