+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "trailingComma": "es5",
+  "semi": true,
+  "singleQuote": true,
+  "tabWidth": 2
+}
