+++ README.md
@@ -21,6 +21,24 @@
 ## Структура папок и файлов
 
 ```
+├── gulpfile.babel.js/            # Папка с настройками Gulp
+│   ├── config                    # Настройки Gulp
+│   │   └── paths.js              # Пути к папкам и файлам
+│   ├── tasks                     # Задачи Gulp
+│   │   ├── cache-task.js         # Задача для кэширования файлов
+│   │   ├── clean.js              # Задача для очистки папки сборки
+│   │   ├── html-include.js       # Задача для сборки HTML
+│   │   ├── images.js             # Задача для обработки изображений
+│   │   ├── index.js              # Входной файл задач
+│   │   ├── resources.js          # Задача для копирования ресурсов
+│   │   ├── rewrite.js            # Задача для перезаписи ссылок на ассеты
+│   │   ├── scripts.js            # Задача для обработки JS
+│   │   ├── scripts-backend.js    # Задача для обработки JS для бэкенда
+│   │   ├── styles.js             # Задача для обработки CSS
+│   │   ├── styles-backend.js     # Задача для обработки CSS для бэкенда
+│   │   ├── svg-sprites.js        # Задача для сборки SVG спрайтов
+│   │   └── zip.js                # Задача для сборки в zip-архив
+│   └── index.js                  # Главный файл Gulp
 ├── src/                          # Исходники
 │   ├── js                        # Скрипты
 │   │   └── main.js               # Главный скрипт
@@ -46,7 +64,6 @@
 │   ├── resources                 # папка для хранения иных ассетов - php, видео-файлы, favicon и т.д.
 │   │   ├── fonts                 # папка для хранения шрифтов в формате woff2
 │   └── index.html                # Главный html-файл
-└── gulpfile.js                   # файл с настройками Gulp
 └── package.json                  # файл с настройками сборки и установленными пакетами
 └── .editorconfig                 # файл с настройками форматирования кода
 └── .ecrc                         # файл с настройками пакета editorconfig-checker (исключает ненужные папки)
