+++ gulpfile.babel.js/config/paths.js
@@ -1,0 +1,23 @@
+import { resolve } from 'path';
+
+// paths
+const rootPath = process.cwd();
+const srcFolder = resolve(rootPath, 'src');
+const buildFolder = resolve(rootPath, 'app');
+const paths = {
+  rootPath,
+  srcFolder,
+  buildFolder,
+  srcSvg: `${srcFolder}/img/svg/**.svg`,
+  srcImgFolder: `${srcFolder}/img`,
+  buildImgFolder: `${buildFolder}/img`,
+  srcScss: `${srcFolder}/scss/**/*.scss`,
+  buildCssFolder: `${buildFolder}/css`,
+  srcFullJs: `${srcFolder}/js/**/*.js`,
+  srcMainJs: `${srcFolder}/js/main.js`,
+  buildJsFolder: `${buildFolder}/js`,
+  srcPartialsFolder: `${srcFolder}/partials`,
+  resourcesFolder: `${srcFolder}/resources`,
+};
+
+export default paths;
