+++ gulpfile.babel.js/tasks/clean.js
@@ -1,0 +1,6 @@
+import del from 'del';
+import paths from '../config/paths';
+
+const clean = () => del([paths.buildFolder]);
+
+export default clean;
