+++ gulpfile.babel.js/tasks/resources.js
@@ -1,0 +1,7 @@
+import { dest, src } from 'gulp';
+import paths from '../config/paths';
+
+const resources = () => src(`${paths.resourcesFolder}/**`)
+    .pipe(dest(paths.buildFolder));
+
+export default resources;
