+++ gulpfile.babel.js/tasks/svg-sprites.js
@@ -1,0 +1,39 @@
+import { dest, src } from 'gulp';
+import svgmin from 'gulp-svgmin';
+import cheerio from 'gulp-cheerio';
+import replace from 'gulp-replace';
+import svgSprite from 'gulp-svg-sprite';
+import paths from '../config/paths';
+
+// svg sprite
+const svgSprites = () => src(paths.srcSvg)
+    .pipe(
+      svgmin({
+        js2svg: {
+          pretty: true,
+        },
+      })
+    )
+    .pipe(
+      cheerio({
+        run ($) {
+          $('[fill]').removeAttr('fill');
+          $('[stroke]').removeAttr('stroke');
+          $('[style]').removeAttr('style');
+        },
+        parserOptions: {
+          xmlMode: true
+        },
+      })
+    )
+    .pipe(replace('&gt;', '>'))
+    .pipe(svgSprite({
+      mode: {
+        stack: {
+          sprite: '../sprite.svg'
+        }
+      },
+    }))
+    .pipe(dest(paths.buildImgFolder));
+
+export default svgSprites;
