+++ package.json
@@ -2,11 +2,11 @@
   "name": "gulp-builder",
   "version": "2.0.4",
   "description": "gulp-builder for weblayout",
-  "main": "gulpfile.js",
+  "main": "gulpfile.babel.js",
   "scripts": {
     "html": "node-w3c-validator -f lint -evH -i app/**/*.html",
     "dev": "gulp",
-    "build": "gulp build",
+    "build": "gulp --production",
     "cache": "gulp cache",
     "backend": "gulp backend",
     "zip": "gulp zip",
@@ -14,13 +14,29 @@
   },
   "author": "MaxGraph",
   "license": "ISC",
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
+  },
+  "browserslist": [
+    "last 5 versions",
+    "> 0.5%",
+    "not dead",
+    "not ie > 0"
+  ],
   "devDependencies": {
     "@babel/core": "^7.16.7",
     "@babel/preset-env": "^7.16.7",
+    "@babel/register": "^7.16.9",
     "babel-loader": "^8.2.3",
     "browser-sync": "^2.26.14",
     "del": "^6.0.0",
     "editorconfig-checker": "^4.0.1",
+    "eslint": "^8.7.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.4",
     "gulp": "^4.0.2",
     "gulp-autoprefixer": "^7.0.1",
     "gulp-avif": "^1.0.1",
@@ -46,6 +62,7 @@
     "node-w3c-validator": "^2.0.1",
     "postcss": "^8.4.5",
     "postcss-scss": "^4.0.2",
+    "prettier": "^2.5.1",
     "sass": "^1.45.1",
     "stylelint": "^14.2.0",
     "stylelint-config-standard-scss": "^3.0.0",
